requirements.txt
aiohttp==2.3.10
aiohttp-swagger==1.0.5
aiopg==0.14.0
aniso8601==3.0.2
async-timeout==2.0.1
chardet==3.0.4
click==6.7
idna==2.7
idna-ssl==1.0.1
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
multidict==4.3.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyJWT==1.6.4
pytz==2018.4
PyYAML==3.12
six==1.11.0
SQLAlchemy==1.2.8
Werkzeug==0.14.1
yarl==1.2.0
